python/requirements.txt
# GigE Camera Dependencies
harvesters==1.4.2
opencv-python==4.8.1.78
numpy==1.24.3

# LiveKit WebRTC Dependencies
livekit==0.11.1
livekit-api==0.6.0

# Utility Dependencies
pyyaml==6.0.1
python-dotenv==1.0.0

# Optional: Aravis support (comment out if not using Aravis)
# PyGObject>=3.42.0